backend/requirements.txt
fastapi[all]>=0.121.0
langchain>=1.0.3
langchain-community>=0.4.1
langchain-ollama>=1.0.0
langchain-openai>=1.0.2
psycopg2-binary>=2.9.11
pydantic-settings>=2.11.0
python-dotenv>=1.2.1
sqlalchemy>=2.0.44
uvicorn>=0.38.0